refactor: extract codewhisperer tree view setup
Problem:
We want to use the CodeWhisperer tree view in the explorer
but when enabling it the current code will try and enable
the other tree views such as the Resource Explorer

Solution:
Extract the logic for the tree view that we want to work
in both node and browser environments in to its own shared
activate function.

Now we can use that shared activation for both environments.
